refactor: notifythread -> root/notifythread.c
