+++ watchman.h
@@ -751,6 +751,7 @@ void w_root_read_unlock(struct read_locked_watchman_root *locked,
 
 void process_pending_symlink_targets(struct unlocked_watchman_root *unlocked);
 void *run_io_thread(void *arg);
+void *run_notify_thread(void *arg);
 void process_subscriptions(struct write_locked_watchman_root *lock);
 void process_triggers(struct write_locked_watchman_root *lock);
 void consider_age_out(struct write_locked_watchman_root *lock);
@@ -758,6 +759,10 @@ bool consider_reap(struct write_locked_watchman_root *lock);
 void remove_from_file_list(struct write_locked_watchman_root *lock,
                            struct watchman_file *file);
 void free_file_node(w_root_t *root, struct watchman_file *file);
+void w_root_teardown(w_root_t *root);
+bool w_root_init(w_root_t *root, char **errmsg);
+bool remove_root_from_watched(
+    w_root_t *root /* don't care about locked state */);
 
 /* Bob Jenkins' lookup3.c hash function */
 uint32_t w_hash_bytes(const void *key, size_t length, uint32_t initval);
