+++ winbuild/Makefile
@@ -93,6 +93,7 @@ SRCS=\
 	root/file.c       \
 	root/iothread.c       \
 	root/lock.c       \
+	root/notifythread.c       \
 	root/symlink.c       \
 	state.c      \
 	string.c     \
